requirements/dev.txt
autopep8
flake8
ipython
pytest
recordtype
tabula-py
twine
wheel
